move from test/adaptive_load to test/common
